+++ src/ResourceManager/Network/Commands.Network/NetworkWatcher/PacketCapture/GetAzureNetworkWatcherPacketCaptureCommand.cs
@@ -58,6 +58,9 @@ namespace Microsoft.Azure.Commands.Network
         [ValidateNotNullOrEmpty]
         public string PacketCaptureName { get; set; }
 
+        [Parameter(Mandatory = false, HelpMessage = "Run cmdlet in the background")]
+        public SwitchParameter AsJob { get; set; }
+
         public override void Execute()
         {
             base.Execute();