+++ src/ResourceManager/OperationalInsights/Commands.OperationalInsights/Workspaces/NewAzureOperationalInsightsWorkspaceCommand.cs
@@ -51,7 +51,9 @@ namespace Microsoft.Azure.Commands.OperationalInsights
 
         [Parameter(Position = 5, Mandatory = false, ValueFromPipelineByPropertyName = true,
             HelpMessage = "The resource tags for the workspace.")]
-        public Hashtable Tags { get; set; }
+        [Obsolete("New-AzureRmOperationalInsightsWorkspace: -Tags will be removed in favor of -Tag in an upcoming breaking change release.  Please start using the -Tag parameter to avoid breaking scripts.")]
+        [Alias("Tags")]
+        public Hashtable Tag { get; set; }
 
         [Parameter(Position = 6, Mandatory = false, ValueFromPipelineByPropertyName = true,
             HelpMessage = "The workspace data retention in days. 730 days is the maximum allowed for all other Skus.")]
@@ -63,6 +65,7 @@ namespace Microsoft.Azure.Commands.OperationalInsights
 
         public override void ExecuteCmdlet()
         {
+#pragma warning disable CS0618
             CreatePSWorkspaceParameters parameters = new CreatePSWorkspaceParameters()
             {
                 ResourceGroupName = ResourceGroupName,
@@ -70,11 +73,12 @@ namespace Microsoft.Azure.Commands.OperationalInsights
                 Location = Location,
                 Sku = Sku,
                 CustomerId = CustomerId,
-                Tags = Tags,
+                Tags = Tag,
                 RetentionInDays = RetentionInDays,
                 Force = Force.IsPresent,
                 ConfirmAction = ConfirmAction
             };
+#pragma warning restore CS0618
 
             WriteObject(OperationalInsightsClient.CreatePSWorkspace(parameters));
         }