+++ src/ResourceManager/Resources/Commands.ResourceManager/Cmdlets/Implementation/CmdletBase/ResourceManagerCmdletBase.cs
@@ -293,13 +293,13 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
             {
                 if (this.resourceManagerSdkClient == null)
                 {
-                    this.resourceManagerSdkClient = new ResourceManagerSdkClient(DefaultContext)
-                    {
-                        VerboseLogger = WriteVerboseWithTimestamp,
-                        ErrorLogger = WriteErrorWithTimestamp,
-                        WarningLogger = WriteWarningWithTimestamp
-                    };
+                    this.resourceManagerSdkClient = new ResourceManagerSdkClient(DefaultContext);
                 }
+
+                this.resourceManagerSdkClient.VerboseLogger = WriteVerboseWithTimestamp;
+                this.resourceManagerSdkClient.ErrorLogger = WriteErrorWithTimestamp;
+                this.resourceManagerSdkClient.WarningLogger = WriteWarningWithTimestamp;
+
                 return this.resourceManagerSdkClient;
             }