+++ src/ResourceManager/Storage/Stack/Commands.Management.Storage.Test/ScenarioTests/StorageAccountTests.ps1
@@ -26,84 +26,62 @@ function Test-StorageAccount
         # Test
         $stoname = 'sto' + $rgname;
         $stotype = 'Standard_GRS';
-        $loc = Get-ProviderLocation ResourceManagement;
-		$encryptionServiceBF = "Blob,File"
-		$encryptionServiceB = "Blob"
-		$encryptionServiceF = "File"
-		$kind = 'BlobStorage'
-		$accessTier = 'Cool'
+        $loc = 'West US';
 
         New-AzureRmResourceGroup -Name $rgname -Location $loc;
 
-        New-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -Location $loc -Type $stotype -Kind $kind -AccessTier $accessTier -EnableEncryptionService $encryptionServiceBF;
+        New-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -Location $loc -Type $stotype;
         $stos = Get-AzureRmStorageAccount -ResourceGroupName $rgname;
 
         $stotype = 'StandardGRS';
         Retry-IfException { $global:sto = Get-AzureRmStorageAccount -ResourceGroupName $rgname  -Name $stoname; }
         Assert-AreEqual $sto.StorageAccountName $stoname;
-        Assert-AreEqual $sto.Sku.Name $stotype;
+        Assert-AreEqual $sto.AccountType $stotype;
         Assert-AreEqual $sto.Location $loc;
-        Assert-AreEqual $sto.Kind $kind;
-        Assert-AreEqual $sto.AccessTier $accessTier;
-		Assert-AreEqual $sto.Encryption.Services.Blob.Enabled $true
-		Assert-AreEqual $sto.Encryption.Services.File.Enabled $true
 
         $stotype = 'Standard_LRS';
-		$accessTier = 'Hot'
         # TODO: Still need to do retry for Set-, even after Get- returns it.
-        Retry-IfException { $global:sto = Set-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -Type $stotype -DisableEncryptionService $encryptionServiceB -AccessTier $accessTier -Force; }
+        Retry-IfException { $global:sto = Set-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -Type $stotype; }
         $sto = Get-AzureRmStorageAccount -ResourceGroupName $rgname  -Name $stoname;
         $stotype = 'StandardLRS';
         Assert-AreEqual $sto.StorageAccountName $stoname;
-        Assert-AreEqual $sto.Sku.Name $stotype;
+        Assert-AreEqual $sto.AccountType $stotype;
         Assert-AreEqual $sto.Location $loc;
-        Assert-AreEqual $sto.Kind $kind;
-        Assert-AreEqual $sto.AccessTier $accessTier;
-		Assert-AreEqual $sto.Encryption.Services.Blob $null
-		Assert-AreEqual $sto.Encryption.Services.File.Enabled $true
     
         $stotype = 'Standard_RAGRS';
-		$accessTier = 'Cool'
-        Set-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -Type $stotype -AccessTier $accessTier -DisableEncryptionService $encryptionServiceF -EnableEncryptionService $encryptionServiceB -Force ;
+        Set-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -Type $stotype;
         
         $sto = Get-AzureRmStorageAccount -ResourceGroupName $rgname  -Name $stoname;
         $stotype = 'StandardRAGRS';
         Assert-AreEqual $sto.StorageAccountName $stoname;
-        Assert-AreEqual $sto.Sku.Name $stotype;
+        Assert-AreEqual $sto.AccountType $stotype;
         Assert-AreEqual $sto.Location $loc;
-        Assert-AreEqual $sto.Kind $kind;
-        Assert-AreEqual $sto.AccessTier $accessTier;
-		Assert-AreEqual $sto.Encryption.Services.Blob.Enabled $true
-		Assert-AreEqual $sto.Encryption.Services.File $null
 
         $stotype = 'Standard_GRS';
-        Set-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -Type $stotype -DisableEncryptionService $encryptionServiceBF;
+        Set-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -Type $stotype;
         
         $sto = Get-AzureRmStorageAccount -ResourceGroupName $rgname  -Name $stoname;
         $stotype = 'StandardGRS';
         Assert-AreEqual $sto.StorageAccountName $stoname;
-        Assert-AreEqual $sto.Sku.Name $stotype;
+        Assert-AreEqual $sto.AccountType $stotype;
         Assert-AreEqual $sto.Location $loc;
-        Assert-AreEqual $sto.Kind $kind;
-        Assert-AreEqual $sto.AccessTier $accessTier;
-		Assert-AreEqual $sto.Encryption $null
 
         $stokey1 = Get-AzureRmStorageAccountKey -ResourceGroupName $rgname -Name $stoname;
 
         New-AzureRmStorageAccountKey -ResourceGroupName $rgname -Name $stoname -KeyName key1;
         
         $stokey2 = Get-AzureRmStorageAccountKey -ResourceGroupName $rgname -Name $stoname;
-        Assert-AreNotEqual $stokey1[0].Value $stokey2[0].Value;
-        Assert-AreEqual $stokey1[1].Value $stokey2[1].Value;
+        Assert-AreNotEqual $stokey1.Key1 $stokey2.Key1;
+        Assert-AreEqual $stokey1.Key2 $stokey2.Key2;
 
         New-AzureRmStorageAccountKey -ResourceGroupName $rgname -Name $stoname -KeyName key2;
 
         $stokey3 = Get-AzureRmStorageAccountKey -ResourceGroupName $rgname -Name $stoname;
-        Assert-AreNotEqual $stokey1[0].Value $stokey2[0].Value;
-        Assert-AreEqual $stokey2[0].Value $stokey3[0].Value;
-        Assert-AreNotEqual $stokey2[1].Value $stokey3[1].Value;
+        Assert-AreNotEqual $stokey1.Key1 $stokey2.Key1;
+        Assert-AreEqual $stokey2.Key1 $stokey3.Key1;
+        Assert-AreNotEqual $stokey2.Key2 $stokey3.Key2;
 
-        Remove-AzureRmStorageAccount -Force -ResourceGroupName $rgname -Name $stoname;
+        Remove-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname;
     }
     finally
     {
@@ -126,13 +104,13 @@ function Test-NewAzureStorageAccount
         # Test
         $stoname = 'sto' + $rgname;
         $stotype = 'Standard_GRS';
-        $loc = Get-ProviderLocation ResourceManagement;
+        $loc = 'West US';
 
         New-AzureRmResourceGroup -Name $rgname -Location $loc;
 
         New-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -Location $loc -Type $stotype;
         
-        Retry-IfException { Remove-AzureRmStorageAccount -Force -ResourceGroupName $rgname -Name $stoname; }
+        Retry-IfException { Remove-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname; }
     }
     finally
     {
@@ -155,29 +133,24 @@ function Test-GetAzureStorageAccount
         # Test
         $stoname = 'sto' + $rgname;
         $stotype = 'Standard_GRS';
-        $loc = Get-ProviderLocation ResourceManagement;
-		$kind = 'Storage'
+        $loc = 'West US';
 
         New-AzureRmResourceGroup -Name $rgname -Location $loc;
 
-        New-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -Location $loc -Type $stotype ;
+        New-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -Location $loc -Type $stotype;
 
         Retry-IfException { $global:sto = Get-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname; }
         $stotype = 'StandardGRS';
         Assert-AreEqual $sto.StorageAccountName $stoname;
-        Assert-AreEqual $sto.Sku.Name $stotype;
+        Assert-AreEqual $sto.AccountType $stotype;
         Assert-AreEqual $sto.Location $loc;
-        Assert-AreEqual $sto.Kind $kind;
-        Assert-AreEqual $sto.EnableHttpsTrafficOnly $false;
 
         $stos = Get-AzureRmStorageAccount -ResourceGroupName $rgname;
         Assert-AreEqual $stos[0].StorageAccountName $stoname;
-        Assert-AreEqual $stos[0].Sku.Name $stotype;
+        Assert-AreEqual $stos[0].AccountType $stotype;
         Assert-AreEqual $stos[0].Location $loc;
-        Assert-AreEqual $sto.Kind $kind;
-        Assert-AreEqual $sto.EnableHttpsTrafficOnly $false;
 
-        Remove-AzureRmStorageAccount -Force -ResourceGroupName $rgname -Name $stoname;
+        Remove-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname;
     }
     finally
     {
@@ -200,48 +173,35 @@ function Test-SetAzureStorageAccount
         # Test
         $stoname = 'sto' + $rgname;
         $stotype = 'Standard_GRS';
-        $loc = Get-ProviderLocation ResourceManagement;
-		$kind = 'Storage'
-		$encryptionServiceBF = "File,Blob"
+        $loc = 'West US';
 
         New-AzureRmResourceGroup -Name $rgname -Location $loc;
-        New-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -Location $loc -Type $stotype -Kind $kind -EnableEncryptionService $encryptionServiceBF -EnableHttpsTrafficOnly $true;
+        New-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -Location $loc -Type $stotype;
 
         Retry-IfException { $global:sto = Get-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname; }
         $stotype = 'StandardGRS';
         Assert-AreEqual $sto.StorageAccountName $stoname;
-        Assert-AreEqual $sto.Sku.Name $stotype;
+        Assert-AreEqual $sto.AccountType $stotype;
         Assert-AreEqual $sto.Location $loc;
-        Assert-AreEqual $sto.Kind $kind;
-        Assert-AreEqual $sto.EnableHttpsTrafficOnly $true;
-		Assert-AreEqual $sto.Encryption.Services.Blob.Enabled $true
-		Assert-AreEqual $sto.Encryption.Services.File.Enabled $true
         
         $stos = Get-AzureRmStorageAccount -ResourceGroupName $rgname;
         Assert-AreEqual $stos[0].StorageAccountName $stoname;
-        Assert-AreEqual $stos[0].Sku.Name $stotype;
+        Assert-AreEqual $stos[0].AccountType $stotype;
         Assert-AreEqual $stos[0].Location $loc;
-        Assert-AreEqual $sto.Kind $kind;
-        Assert-AreEqual $sto.EnableHttpsTrafficOnly $true;
-		Assert-AreEqual $sto.Encryption.Services.Blob.Enabled $true
-		Assert-AreEqual $sto.Encryption.Services.File.Enabled $true
 
         $stotype = 'Standard_LRS';
         # TODO: Still need to do retry for Set-, even after Get- returns it.
-        Retry-IfException { Set-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -Type $stotype -DisableEncryptionService $encryptionServiceBF -EnableHttpsTrafficOnly $false; }
+        Retry-IfException { Set-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -Type $stotype; }
         $stotype = 'Standard_RAGRS';
         Set-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -Type $stotype;
 
         $sto = Get-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname;
         $stotype = 'StandardRAGRS';
         Assert-AreEqual $sto.StorageAccountName $stoname;
-        Assert-AreEqual $sto.Sku.Name $stotype;
+        Assert-AreEqual $sto.AccountType $stotype;
         Assert-AreEqual $sto.Location $loc;
-        Assert-AreEqual $sto.Kind $kind;
-        Assert-AreEqual $sto.EnableHttpsTrafficOnly $false;
-		Assert-AreEqual $sto.Encryption $null
 
-        Remove-AzureRmStorageAccount -Force -ResourceGroupName $rgname -Name $stoname;
+        Remove-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname;
     }
     finally
     {
@@ -252,7 +212,7 @@ function Test-SetAzureStorageAccount
 
 <#
 .SYNOPSIS
-Test Remove-AzureRmStorageAccount -Force
+Test Remove-AzureRmStorageAccount
 #>
 function Test-RemoveAzureStorageAccount
 {
@@ -264,65 +224,13 @@ function Test-RemoveAzureStorageAccount
         # Test
         $stoname = 'sto' + $rgname;
         $stotype = 'Standard_GRS';
-        $loc = Get-ProviderLocation ResourceManagement;
+        $loc = 'West US';
 
         New-AzureRmResourceGroup -Name $rgname -Location $loc;
 
         New-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -Location $loc -Type $stotype;
         
-        Retry-IfException { Remove-AzureRmStorageAccount -Force -ResourceGroupName $rgname -Name $stoname; }
-    }
-    finally
-    {
-        # Cleanup
-        Clean-ResourceGroup $rgname
-    }
-}
-
-<#
-.SYNOPSIS
-Test New-AzureRmStorageAccountEncryptionKeySource
-#>
-function Test-SetAzureRmStorageAccountKeySource
-{
-    # Setup
-    $rgname = Get-StorageManagementTestResourceName;
-
-    try
-    {
-        # Test
-        $stoname = 'sto' + $rgname;
-        $stotype = 'Standard_GRS';
-        $loc = Get-ProviderLocation ResourceManagement;
-
-        New-AzureRmResourceGroup -Name $rgname -Location $loc;
-
-        New-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -Location $loc -Type $stotype;
-		
-		$sto = Set-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -EnableEncryptionService Blob -StorageEncryption
-		$stotype = 'StandardGRS';
-        Assert-AreEqual $sto.StorageAccountName $stoname;
-        Assert-AreEqual $sto.Sku.Name $stotype;
-        Assert-AreEqual $sto.Location $loc;
-		Assert-AreEqual $sto.Encryption.Services.Blob.Enabled $true
-        Assert-AreEqual $sto.Encryption.KeySource Microsoft.Storage;
-        Assert-AreEqual $sto.Encryption.Keyvaultproperties.Keyname $null;
-        Assert-AreEqual $sto.Encryption.Keyvaultproperties.KeyVersion $null;
-        Assert-AreEqual $sto.Encryption.Keyvaultproperties.KeyVaultUri $null;
-		
-		$sto = Set-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -EnableEncryptionService File -AssignIdentity
-        Assert-AreEqual $sto.StorageAccountName $stoname;
-        Assert-AreEqual $sto.Sku.Name $stotype;
-        Assert-AreEqual $sto.Location $loc;
-		Assert-AreNotEqual $sto.Identity.Type SystemAssigned $null
-		Assert-AreEqual $sto.Encryption.Services.Blob.Enabled $true
-		Assert-AreEqual $sto.Encryption.Services.File.Enabled $true
-        Assert-AreEqual $sto.Encryption.KeySource Microsoft.Storage;
-        Assert-AreEqual $sto.Encryption.Keyvaultproperties.Keyname $null;
-        Assert-AreEqual $sto.Encryption.Keyvaultproperties.KeyVersion $null;
-        Assert-AreEqual $sto.Encryption.Keyvaultproperties.KeyVaultUri $null;
-		
-        Remove-AzureRmStorageAccount -Force -ResourceGroupName $rgname -Name $stoname; 
+        Retry-IfException { Remove-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname; }
     }
     finally
     {
@@ -345,16 +253,16 @@ function Test-GetAzureStorageAccountKey
         # Test
         $stoname = 'sto' + $rgname;
         $stotype = 'Standard_GRS';
-        $loc = Get-ProviderLocation ResourceManagement;
+        $loc = 'West US';
 
         New-AzureRmResourceGroup -Name $rgname -Location $loc;
 
         New-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -Location $loc -Type $stotype;
         
         Retry-IfException { $global:stokeys = Get-AzureRmStorageAccountKey -ResourceGroupName $rgname -Name $stoname; }
-        Assert-AreNotEqual $stokeys[0].Value $stokeys[1].Value;
+        Assert-AreNotEqual $stokeys.Key1 $stokeys.Key2;
 
-        Remove-AzureRmStorageAccount -Force -ResourceGroupName $rgname -Name $stoname;
+        Remove-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname;
     }
     finally
     {
@@ -377,7 +285,7 @@ function Test-NewAzureStorageAccountKey
         # Test
         $stoname = 'sto' + $rgname;
         $stotype = 'Standard_GRS';
-        $loc = Get-ProviderLocation ResourceManagement;
+        $loc = 'West US';
 
         New-AzureRmResourceGroup -Name $rgname -Location $loc;
 
@@ -388,17 +296,17 @@ function Test-NewAzureStorageAccountKey
         New-AzureRmStorageAccountKey -ResourceGroupName $rgname -Name $stoname -KeyName key1;
 
         $stokey2 = Get-AzureRmStorageAccountKey -ResourceGroupName $rgname -Name $stoname;
-        Assert-AreNotEqual $stokey1[0].Value $stokey2[0].Value;
-        Assert-AreEqual $stokey1[1].Value $stokey2[1].Value;
+        Assert-AreNotEqual $stokey1.Key1 $stokey2.Key1;
+        Assert-AreEqual $stokey1.Key2 $stokey2.Key2;
 
         New-AzureRmStorageAccountKey -ResourceGroupName $rgname -Name $stoname -KeyName key2;
 
         $stokey3 = Get-AzureRmStorageAccountKey -ResourceGroupName $rgname -Name $stoname;
-        Assert-AreNotEqual $stokey1[0].Value $stokey2[0].Value;
-        Assert-AreEqual $stokey2[0].Value $stokey3[0].Value;
-        Assert-AreNotEqual $stokey2[1].Value $stokey3[1].Value;
+        Assert-AreNotEqual $stokey1.Key1 $stokey2.Key1;
+        Assert-AreEqual $stokey2.Key1 $stokey3.Key1;
+        Assert-AreNotEqual $stokey2.Key2 $stokey3.Key2;
 
-        Remove-AzureRmStorageAccount -Force -ResourceGroupName $rgname -Name $stoname;
+        Remove-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname;
     }
     finally
     {
@@ -421,16 +329,16 @@ function Test-PipingGetAccountToGetKey
         # Test
         $stoname = 'sto' + $rgname;
         $stotype = 'Standard_GRS';
-        $loc = Get-ProviderLocation ResourceManagement;
+        $loc = 'West US';
 
         New-AzureRmResourceGroup -Name $rgname -Location $loc;
 
         New-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -Location $loc -Type $stotype;
 
         Retry-IfException { $global:stokeys = Get-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname | Get-AzureRmStorageAccountKey -ResourceGroupName $rgname; }
-        Assert-AreNotEqual $stokeys[0].Value $stokeys[1].Value;
+        Assert-AreNotEqual $stokeys.Key1 $stokeys.Key2;
 
-        Remove-AzureRmStorageAccount -Force -ResourceGroupName $rgname -Name $stoname;
+        Remove-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname;
     }
     finally
     {
@@ -453,16 +361,15 @@ function Test-PipingToSetAzureRmCurrentStorageAccount
         # Test
         $stoname = 'sto' + $rgname
         $stotype = 'Standard_GRS'
-        $loc = Get-ProviderLocation ResourceManagement
+        $loc = 'West US'
 
         New-AzureRmResourceGroup -Name $rgname -Location $loc
         New-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -Location $loc -Type $stotype
         Retry-IfException { $global:sto = Get-AzureRmStorageAccount -ResourceGroupName $rgname  -Name $stoname }
 		$global:sto | Set-AzureRmCurrentStorageAccount
 		$context = Get-AzureRmContext
-		$sub = New-Object -TypeName Microsoft.Azure.Commands.Profile.Models.PSAzureSubscription -ArgumentList $context.Subscription
-		Assert-AreEqual $stoname $sub.CurrentStorageAccountName
-		$global:sto | Remove-AzureRmStorageAccount -Force
+		Assert-AreEqual $stoname $context.Subscription.CurrentStorageAccountName
+		$global:sto | Remove-AzureRmStorageAccount
     }
     finally
     {
@@ -485,7 +392,7 @@ function Test-SetAzureRmCurrentStorageAccount
         # Test
         $stoname = 'sto' + $rgname
         $stotype = 'Standard_GRS'
-        $loc = Get-ProviderLocation ResourceManagement
+        $loc = 'West US'
 
         New-AzureRmResourceGroup -Name $rgname -Location $loc
         New-AzureRmStorageAccount -ResourceGroupName $rgname -Name $stoname -Location $loc -Type $stotype
@@ -493,7 +400,7 @@ function Test-SetAzureRmCurrentStorageAccount
 		Set-AzureRmCurrentStorageAccount -ResourceGroupName $rgname -StorageAccountName $stoname
 		$context = Get-AzureRmContext
 		Assert-AreEqual $stoname $context.Subscription.CurrentStorageAccountName
-		$global:sto | Remove-AzureRmStorageAccount -Force
+		$global:sto | Remove-AzureRmStorageAccount
     }
     finally
     {