+++ src/ResourceManager/AnalysisServices/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Added ResourceGroup Completer to -ResourceGroup parameters allowing tab completion through resource groups in current subscription
 
 ## Version 0.6.1
 * Remove validate set of location into dynamic lookup so that all clouds are supported.